Add files via upload
The first eight files contain the Microsoft excel documents for each sheet of the analysis. All other documents downward are the pictures of the analysis. Some of the analysis were also done in Power Bi but later copied on the microsoft excel sheets because Excel has more rooms for larger dashboards.